Renaming, simplify date test setup
